refactor: move the VIP-254 into its own folder
